feat(crossplatform): add cross platform support
This adds support for the role to work on arm linux, arm64 linux, and
amd64 darwin. It attempts to determine the appropraite architecture and
operating system choice and download the appropriate binary and check
the SHA256 of that downloaded binary.

In addition, a small change had to be made to the creation of the ctop
installation directory for operating systems where `/usr/local/bin` may
not be owned by root, like MacOS with homebrew.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>